refactor import error list
