--- packages/compass-import-export/src/components/import-modal/import-modal.module.less
@@ -1,25 +1,0 @@
-@import (reference) 'mongodb-compass/src/app/styles/index.less';
-
-.import-modal {
-  &-errors {
-    margin-top: 10px;
-  }
-
-  &-errors-list {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-  }
-
-  &-errors-list-item:not(:first-child) {
-    margin-top: 10px;
-  }
-
-  &-show-more-errors-button {
-    margin-top: 10px;
-
-    &-text {
-      margin-left: 5px;
-    }
-  }
-}
